refactor: rename file
ecalc.py is actually energy calculator
