+++ src/libecalc/core/energy_calculator.py
