+++ src/tests/libecalc/dto/results/model/test_requested_pressures_compressors.py
@@ -4,7 +4,7 @@ from typing import List
 import pytest
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import TimeSeriesFloat
-from libecalc.core.ecalc import EnergyCalculator
+from libecalc.core.energy_calculator import EnergyCalculator
 from libecalc.core.graph_result import EcalcModelResult, GraphResult
 from libecalc.dto.result.results import CompressorModelResult
