+++ src/tests/libecalc/integration/test_all_energy_usage_models.py
@@ -1,5 +1,5 @@
 import pytest
-from libecalc.core.ecalc import EnergyCalculator
+from libecalc.core.energy_calculator import EnergyCalculator
 from libecalc.core.graph_result import EnergyCalculatorResult, GraphResult
