+++ src/tests/libecalc/integration/test_multiple_installations.py
@@ -4,7 +4,7 @@ import pytest
 from libecalc import dto
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import RateType, TimeSeriesRate
-from libecalc.core.ecalc import EnergyCalculator
+from libecalc.core.energy_calculator import EnergyCalculator
 from libecalc.core.graph_result import GraphResult
